+++ docs/docs/ProgrammingGuide/AutoML/forecasting.md
@@ -13,7 +13,7 @@ _TimeSequencePredictor_ can be used to train a model on historical time sequence
      
 ```python
 from zoo import init_spark_on_local
-from zoo.ray.util.raycontext import RayContext
+from zoo.ray import RayContext
 sc = init_spark_on_local(cores=4)
 ray_ctx = RayContext(sc=sc)
 ray_ctx.init()
@@ -23,7 +23,7 @@ ray_ctx.init()
       
 ```python
 from zoo import init_spark_on_yarn
-from zoo.ray.util.raycontext import RayContext
+from zoo.ray import RayContext
 slave_num = 2
 sc = init_spark_on_yarn(
         hadoop_conf=args.hadoop_conf,
