+++ docs/docs/Zouwu/tutorial.md
@@ -85,7 +85,7 @@ You'll need ```RayOnSpark``` for training with ```AutoTSTrainer```, so you have 
 
 ```python
 from zoo import init_spark_on_local
-from zoo.ray.util.raycontext import RayContext
+from zoo.ray import RayContext
 sc = init_spark_on_local(cores=4)
 ray_ctx = RayContext(sc=sc)
 ray_ctx.init()
@@ -94,7 +94,7 @@ ray_ctx.init()
 
 ```python
 from zoo import init_spark_on_yarn
-from zoo.ray.util.raycontext import RayContext
+from zoo.ray import RayContext
 slave_num = 2
 sc = init_spark_on_yarn(
         hadoop_conf=args.hadoop_conf,
