+++ pyzoo/test/zoo/automl/conftest.py
@@ -21,7 +21,7 @@ ray_ctx = None
 @pytest.fixture(autouse=True, scope='package')
 def automl_fixture():
     from zoo import init_spark_on_local
-    from zoo.ray.util.raycontext import RayContext
+    from zoo.ray import RayContext
     sc = init_spark_on_local(cores=4, spark_log_level="INFO")
     ray_ctx = RayContext(sc=sc, object_store_memory="1g")
     ray_ctx.init()
