+++ pyzoo/test/zoo/ray/test_reinit_raycontext.py
@@ -13,16 +13,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+import time
 from unittest import TestCase
 
 import numpy as np
 import psutil
 import pytest
 import ray
-import time
 
 from zoo import init_spark_on_local
-from zoo.ray.util.raycontext import RayContext
+from zoo.ray import RayContext
 
 np.random.seed(1337)  # for reproducibility
