+++ pyzoo/zoo/examples/ray/rl_pong/rl_pong.py
@@ -20,15 +20,17 @@
 from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
-import os
+
 import argparse
-import numpy as np
 import os
-import ray
 import time
+
 import gym
+import numpy as np
+import ray
+
 from zoo import init_spark_on_yarn, init_spark_on_local
-from zoo.ray.util.raycontext import RayContext
+from zoo.ray import RayContext
 
 os.environ["LANG"] = "C.UTF-8"
 # Define some hyperparameters.
