+++ pyzoo/zoo/ray/__init__.py
@@ -13,3 +13,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+
+from .raycontext import RayContext
