+++ pyzoo/zoo/ray/utils.py
@@ -14,10 +14,44 @@
 # limitations under the License.
 #
 
+import re
 import os
 import signal
 
-from zoo.ray.util.utils import to_list
+
+def to_list(input):
+    if isinstance(input, (list, tuple)):
+        return list(input)
+    else:
+        return [input]
+
+
+def resource_to_bytes(resource_str):
+    if not resource_str:
+        return resource_str
+    matched = re.compile("([0-9]+)([a-z]+)?").match(resource_str.lower())
+    fraction_matched = re.compile("([0-9]+\\.[0-9]+)([a-z]+)?").match(resource_str.lower())
+    if fraction_matched:
+        raise Exception(
+            "Fractional values are not supported. Input was: {}".format(resource_str))
+    try:
+        value = int(matched.group(1))
+        postfix = matched.group(2)
+        if postfix == 'b':
+            value = value
+        elif postfix == 'k':
+            value = value * 1000
+        elif postfix == "m":
+            value = value * 1000 * 1000
+        elif postfix == 'g':
+            value = value * 1000 * 1000 * 1000
+        else:
+            raise Exception("Not supported type: {}".format(resource_str))
+        return value
+    except Exception:
+        raise Exception("Size must be specified as bytes(b),"
+                        "kilobytes(k), megabytes(m), gigabytes(g). "
+                        "E.g. 50b, 100k, 250m, 30g")
 
 
 def gen_shutdown_per_node(pgids, node_ips=None):
@@ -42,5 +76,5 @@ def gen_shutdown_per_node(pgids, node_ips=None):
 
 
 def is_local(sc):
-    master = sc._conf.get("spark.master")
+    master = sc.getConf().get("spark.master")
     return master == "local" or master.startswith("local[")
